requirements.txt
Pillow==10.1.0
rich==13.6.0
requests==2.31.0
typer==0.9.0
